+++ src/utils.ts
@@ -1,30 +1,74 @@
-import childProcess from 'child_process';
-import fetch, { Response } from 'node-fetch';
-import { titleCase } from './case-utils';
-import { ClickUp } from './clickup';
-import { CONFIG } from './config';
-import { ChecklistItem } from './models/clickup.models';
-import { Issue } from './models/gitlab/issue.models';
-import { HttpMethod, NormalizedChecklist, Site } from './models/models';
+import childProcess from "child_process";
+import fetch, { RequestInfo, RequestInit, Response } from "node-fetch";
+import { titleCase } from "./case-utils";
+import { ClickUp } from "./clickup";
+import { CONFIG } from "./config";
+import { ChecklistItem } from "./models/clickup.models";
+import { Issue } from "./models/gitlab/issue.models";
+import { HttpMethod, NormalizedChecklist, Site } from "./models/models";
 
-function checkStatus(res: Response) {
-  if (res.ok) {
-    return res;
+const RETRY_SETTING = {
+  retry: 5,
+  pause: 12 * 1000,
+};
+
+function sleep(ms: number) {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+}
+
+async function fetchRetry(
+  url: RequestInfo,
+  opts?: RequestInit & {
+    retry?: number;
+    callback?: (retry?: number) => any;
+    pause?: number;
+    silent?: boolean;
+  }
+) {
+  let retry = (opts && opts.retry) || 3;
+  while (retry > 0) {
+    try {
+      return await fetch(url, opts);
+    } catch (e) {
+      if (opts?.callback) {
+        opts.callback(retry);
+      }
+      retry = retry - 1;
+      if (retry == 0) {
+        throw e;
+      }
+
+      if (opts?.pause) {
+        if (!opts?.silent) console.log("pausing..");
+        await sleep(opts.pause);
+        if (!opts?.silent) console.log("done pausing...");
+      }
+    }
+  }
+}
+
+function checkStatus(res: Response | undefined) {
+  if (res) {
+    if (res.ok) {
+      return res;
+    } else {
+      throw Error(res.statusText);
+    }
   } else {
-    throw Error(res.statusText);
+    throw Error("Response is undefined.");
   }
 }
 
 export function callApiFactory(site: Site) {
-  let apiUrl = '';
+  let apiUrl = "";
   let headers = {};
   switch (site) {
-    case 'GitLab':
-      apiUrl = 'https://gitlab.com/api/v4';
-      headers = { 'Private-Token': CONFIG.GitLabToken };
+    case "GitLab":
+      apiUrl = "https://gitlab.com/api/v4";
+      headers = { "Private-Token": CONFIG.GitLabToken };
       break;
-    case 'ClickUp':
-      apiUrl = 'https://api.clickup.com/api/v2';
+    case "ClickUp":
+      apiUrl = "https://api.clickup.com/api/v2";
       headers = { Authorization: CONFIG.ClickUpToken };
       break;
     default:
@@ -41,14 +85,15 @@ export function callApiFactory(site: Site) {
         params.set(key, value);
       });
     }
-    return fetch(
+    return fetchRetry(
       apiUrl + url,
-      method === 'get'
+      method === "get"
         ? {
             method,
             headers,
+            ...RETRY_SETTING,
           }
-        : { method, headers, body: params }
+        : { method, headers, body: params, ...RETRY_SETTING }
     )
       .then(checkStatus)
       .then((res) => res.json());
@@ -57,14 +102,14 @@ export function callApiFactory(site: Site) {
 
 export function dashify(input: string) {
   let temp = input
-    .replace(/[^A-Za-z0-9]/g, '-')
-    .replace(/-{2,}/g, '-')
-    .replace(/-+$/, '')
-    .replace(/^-+/, '')
+    .replace(/[^A-Za-z0-9]/g, "-")
+    .replace(/-{2,}/g, "-")
+    .replace(/-+$/, "")
+    .replace(/^-+/, "")
     .toLowerCase();
   if (temp.length >= 100) {
     temp = temp.substring(0, 100);
-    return temp.substring(0, temp.lastIndexOf('-'));
+    return temp.substring(0, temp.lastIndexOf("-"));
   }
   return temp;
 }
@@ -73,14 +118,14 @@ export function normalizeGitLabIssueChecklist(
   checklistText: string
 ): NormalizedChecklist {
   return checklistText
-    .split('\n')
+    .split("\n")
     .filter(
-      (line) => line && (line.includes('- [ ]') || line.includes('- [x]'))
+      (line) => line && (line.includes("- [ ]") || line.includes("- [x]"))
     )
     .map((line, index) => ({
       name: line
-        .replace(/- \[[x ]\] /g, '')
-        .replace(/^ +/, (space) => space.replace(/ /g, '-')),
+        .replace(/- \[[x ]\] /g, "")
+        .replace(/^ +/, (space) => space.replace(/ /g, "-")),
       checked: /- \[x\]/.test(line),
       order: index,
     }));
@@ -102,8 +147,8 @@ export function normalizeClickUpChecklist(
 export async function promiseSpawn(command: string, args: string[]) {
   return new Promise((resolve, reject) => {
     childProcess
-      .spawn(command, args, { shell: true, stdio: 'inherit' })
-      .on('close', (code) => (code === 0 ? resolve() : reject()));
+      .spawn(command, args, { shell: true, stdio: "inherit" })
+      .on("close", (code) => (code === 0 ? resolve() : reject()));
   });
 }