+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unrealeased] - yyyy-mm-dd
 
 ### Added
+- Add retry to api request
 
 ### Changed